Finished Days 60 and 61
Also updated the directory structure to ease the coding involving Flask. So each project that involves HTML or other files will now be created in a directory with its respective day. There will still be Python files for each day, but for those with directories, the file will indicate the code is placed there.